+++ Sources/StreamChatSwiftUI/ChatChannel/ChannelInfo/PinnedMessagesViewModel.swift
@@ -11,7 +11,8 @@ public class PinnedMessagesViewModel: ObservableObject {
     let channel: ChatChannel
 
     @Published var pinnedMessages: [ChatMessage]
-    
+    @Published var selectedMessage: ChatMessage?
+
     private var channelController: ChatChannelController?
     
     public init(channel: ChatChannel, channelController: ChatChannelController? = nil) {